Reorganise stuff in backend split cell game impl
