+++ minegauler/core/board.py
@@ -26,13 +26,25 @@ class BoardBase(metaclass=abc.ABCMeta):
     def __getitem__(self, key: Coord) -> CellContents:
         raise NotImplementedError
 
-    @property
     @abc.abstractmethod
-    def all_coords(self) -> List:
+    def __setitem__(self, key: Coord, obj: CellContents) -> None:
         raise NotImplementedError
 
+    @property
     @abc.abstractmethod
-    def get_nbrs(self, coord, *, include_origin: bool = False) -> Iterable:
+    def all_coords(self) -> List[Coord]:
+        """A list of all coords currently in the board."""
+        raise NotImplementedError
+
+    @property
+    def all_underlying_coords(self) -> List[Coord]:
+        """A list of all underlying coords that may contain a mine."""
+        return self.all_coords
+
+    @abc.abstractmethod
+    def get_nbrs(
+        self, coord: Coord, *, include_origin: bool = False
+    ) -> Iterable[Coord]:
         raise NotImplementedError
 
     @abc.abstractmethod
